packages/bytebot-holo/pyproject.toml
[project]
name = "bytebot-omniparser"
version = "1.0.0"
description = "OmniParser v2.0 integration for Bytebot Hawkeye CV pipeline"
authors = [
    {name = "Bytebot Hawkeye Team"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "huggingface-hub>=0.26.2",
    "ultralytics>=8.3.0",
    "transformers>=4.46.0",
    "opencv-python>=4.10.0",
    "Pillow>=10.4.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "einops>=0.8.0",
    "timm>=1.0.11",
    "numpy>=1.24.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.scripts]
omniparser-server = "src.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
